gate/pyproject.toml
[project]
name = "gate"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "itsdangerous>=2.2",
    "httpx>=0.28",
    "pyyaml>=6.0",
    "python-multipart>=0.0.18",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
